Test method D (#753)
Created tests for validating the calculation of the classical Voronoi data and IFD data with method D. Tests for validating the cut off of voronoi cells and geometry intersection are also provided.
Tests are calculating with artificial trajectories which are a grid that is moving with constant speed through a corridor. Python script for creating them is provided.