Events (#789)
* Remove unused method

* Decouple EventManager from Building

EventManager no longer depends on Building. If an event needs access to
Building is depndent only on what the event will do. Concrete events now
need a building as construction parameter if they want to use it.

This makes sure that all events are fully initialised on construction.

* Make Events ostream compatible

This change allows Events and all subtypes to be printed with libfmt
directly when using the object as parameter.

* Remove EventManager dependency on Pedestrian

Accept the time to which events should be processed as parameter
instead.

* Only sort events when processing events

* Simplified EventManager::ProcessEvents()

* fixup! Decouple EventManager from Building

* Cleanup EventManager.h/cpp

* Remove Event Enum

Events are no longer using a combination of C++Types / Enum tags.