Expressive warning for overlapping pedestrians (#824)
* Expressive warning for overlapping pedestrians

Total overlap of pedestrians in force-based models (gcfm)
can not be fixed in a reliable way.

Here, we give a warning and let the user decide whether the simulation results
can be trusted or not. Normally, "a little" overlapping is acceptable. However,
if the simulation produces too much warnings of this kind, the strength of the
force should be adjusted and the simulation has to be performed with different
parameter values.

* Add note about overlaping in models